+++ notifications/src/main/java/org/opennms/horizon/notifications/grpc/config/GrpcConfig.java
@@ -35,6 +35,8 @@ import org.keycloak.adapters.KeycloakDeployment;
 import org.keycloak.adapters.rotation.JWKPublicKeyLocator;
 import org.keycloak.representations.adapters.config.AdapterConfig;
 import org.opennms.horizon.notifications.grpc.service.NotificationGrpcService;
+import org.opennms.horizon.notifications.tenant.TenantContextLookupImpl;
+import org.opennms.horizon.notifications.tenant.TenantLookup;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -49,11 +51,6 @@ public class GrpcConfig {
     private String keycloakAuthUrl;
     @Value("${horizon.keycloak.realm}")
     private String keycloakRealm;
-
-    @Bean
-    public TenantLookup createTenantLookup(){
-        return new TenantContextLookupImpl();
-    }
 
     @Bean
     public KeycloakDeployment createKeycloak() {
