+++ notifications/src/main/java/org/opennms/horizon/notifications/tenant/TenantLookup.java
@@ -26,7 +26,7 @@
  *     http://www.opennms.com/
  *******************************************************************************/
 
-package org.opennms.horizon.notifications.grpc.config;
+package org.opennms.horizon.notifications.tenant;
 
 import io.grpc.Context;
