+++ notifications/src/main/java/org/opennms/horizon/notifications/tenant/WithTenantAspect.java
@@ -31,7 +31,6 @@ package org.opennms.horizon.notifications.tenant;
 import org.aspectj.lang.ProceedingJoinPoint;
 import org.aspectj.lang.annotation.Around;
 import org.aspectj.lang.annotation.Aspect;
-import org.opennms.horizon.notifications.grpc.config.TenantLookup;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
